Move cloud compliance population to separate job
