+++ deepfence_worker/cronjobs/neo4j.go
@@ -1,9 +1,7 @@
 package cronjobs
 
 import (
-	"encoding/json"
 	"fmt"
-	"os"
 	"time"
 
 	"github.com/ThreeDotsLabs/watermill/message"
@@ -20,30 +18,6 @@ const (
 	dbScanTimeout            = time.Minute * 2
 	dbUpgradeTimeout         = time.Minute * 5
 )
-
-var BenchmarksAvailable = []string{"cis", "nist", "pci", "gdpr", "hipaa", "soc_2"}
-
-type Benchmark struct {
-	BenchmarkId   string            `json:"benchmark_id"`
-	Description   string            `json:"description"`
-	Title         string            `json:"title"`
-	Tags          map[string]string `json:"tags"`
-	Documentation string            `json:"documentation"`
-	Children      []string          `json:"children"`
-}
-
-type Control struct {
-	CategoryBreadcrumb      string            `json:"category_breadcrumb"`
-	CategoryHierarchy       []string          `json:"category_hierarchy"`
-	ControlId               string            `json:"control_id"`
-	Description             string            `json:"description"`
-	Title                   string            `json:"title"`
-	Tags                    map[string]string `json:"tags"`
-	Documentation           string            `json:"documentation"`
-	ParentControlHierarchy  []string          `json:"parent_control_hierarchy"`
-	ParentControlBreadcrumb string            `json:"parent_control_breadcrumb"`
-	Executable              bool              `json:"executable"`
-}
 
 func CleanUpDB(msg *message.Message) error {
 	namespace := msg.Metadata.Get(directory.NamespaceKey)
@@ -241,85 +215,6 @@ func ApplyGraphDBStartup(msg *message.Message) error {
 	addIndexOnIssuesCount(session, "ContainerImage")
 	addIndexOnIssuesCount(session, "Container")
 
-	err = addCloudControls(session)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func addCloudControls(session neo4j.Session) error {
-	cwd := "/cloud_controls/aws"
-	for _, benchmark := range BenchmarksAvailable {
-		controlFilePath := fmt.Sprintf("%s/%s.json", cwd, benchmark)
-		controlsJson, err := os.ReadFile(controlFilePath)
-		if err != nil {
-			return fmt.Errorf("Error reading controls file %s: %s", controlFilePath, err.Error())
-		}
-		var controlList []Control
-		if err := json.Unmarshal(controlsJson, &controlList); err != nil {
-			return fmt.Errorf("Error unmarshalling controls for compliance type %s: %s", benchmark, err.Error())
-		}
-		if _, err = session.Run(`
-		UNWIND $batch as row
-		MERGE (n:CloudComplianceExecutable:CloudComplianceControl{
-			node_id: row.parent_control_breadcrumb + row.control_id,
-			control_id: row.control_id,
-			type: 'control',
-			description: row.description,
-			title: row.title,
-			documentation: row.documentation,
-			service: row.tags.service,
-			cloud_provider: row.tags.plugin,
-			category: row.tags.category,
-			compliance_type: $benchmark,
-			parent_control_hierarchy: row.parent_control_hierarchy,
-			parent_control_breadcrumb: row.parent_control_breadcrumb,
-			category_hierarchy: row.category_hierarchy,
-			category_breadcrumb: row.category_breadcrumb,
-			executable: row.executable
-		})
-		ON CREATE
-			SET n.active = true`,
-			map[string]interface{}{
-				"batch":     controlList,
-				"benchmark": benchmark,
-			}); err != nil {
-			return err
-		}
-		benchmarkFilePath := fmt.Sprintf("%s/%s_benchmarks.json", cwd, benchmark)
-		benchmarksJson, err := os.ReadFile(benchmarkFilePath)
-		if err != nil {
-			return fmt.Errorf("Error reading benchmarks file %s: %s", benchmarkFilePath, err.Error())
-		}
-		var benchmarkList []Benchmark
-		if err := json.Unmarshal(benchmarksJson, &benchmarkList); err != nil {
-			return fmt.Errorf("Error unmarshalling benchmarks for compliance type %s: %s", benchmark, err.Error())
-		}
-		if _, err = session.Run(`
-		UNWIND $batch as row
-		MERGE (n:CloudComplianceExecutable:CloudComplianceBenchmark{
-			node_id: row.benchmark_id,
-			benchmark_id: row.benchmark_id,
-			type: 'benchmark',
-			description: row.description,
-			title: row.title,
-			documentation: row.documentation,
-			service: row.tags.service,
-			cloud_provider: row.tags.plugin,
-			category: row.tags.category,
-			executable: false
-		})
-		UNWIND row.children AS childControl
-		MATCH (m:CloudComplianceExecutable{control_id: childControl})
-		MERGE (n) -[:INCLUDES]-> (m)
-		`,
-			map[string]interface{}{
-				"batch": benchmarkList,
-			}); err != nil {
-			return err
-		}
-	}
 	return nil
 }
